atelier1/dynamic: refactoring packages DTO/DAO
