+++ atelier1/dynamic/src/main/java/cpe/atelier1/controller/RequestCrt.java
@@ -1,16 +1,12 @@
 package cpe.atelier1.controller;
 
-import cpe.atelier1.model.Card;
-import cpe.atelier1.model.CardFormDTO;
+import cpe.atelier1.model.dto.CardFormDTO;
 import cpe.atelier1.services.CardService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.ArrayList;
-import java.util.List;
 
 @Controller
 public class RequestCrt {
@@ -54,5 +50,6 @@ public class RequestCrt {
         model.addAttribute("card", new CardFormDTO());
         return "new/addCard";
     }
+
 
 }
