+++ atelier1/dynamic/src/main/java/cpe/atelier1/services/CardService.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
-import cpe.atelier1.controller.CardDao;
+import cpe.atelier1.model.dao.CardDao;
 import cpe.atelier1.model.Card;
-import cpe.atelier1.model.CardFormDTO;
+import cpe.atelier1.model.dto.CardFormDTO;
 import cpe.atelier1.serialization.CardSerializer;
 import cpe.atelier1.serialization.impl.CardJSONSerializer;
 import org.slf4j.Logger;
